Move getData from index to /lib
